i16 move values

+++ src/movegen.h
@@ -37,7 +37,7 @@ enum GenType {
 };
 
 struct ExtMove: public Move {
-    int value;
+    i16 value;
 
     void operator=(Move m) { data = m.raw(); }
